[implementation] moving local sources implementation to app module
