+++ app/build.gradle
@@ -16,7 +16,7 @@ android {
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
-        resValue "string", "movie_db_api_key", MOVIE_DB_API_KEY
+        buildConfigField "String", "movie_db_api_key", MOVIE_DB_API_KEY
         buildConfigField "String", "BASE_URL", "\"https://api.themoviedb.org/3/\""
     }
     buildTypes {
