+++ app/src/main/java/com/architectcoders/equipocinco/ui/popular/MoviesFragment.kt
@@ -1,4 +1,4 @@
-package com.architectcoders.equipocinco.ui
+package com.architectcoders.equipocinco.ui.popular
 
 import android.Manifest
 import android.os.Bundle
@@ -19,6 +19,8 @@ import com.architectcoders.equipocinco.di.modules.PopularMoviesModule
 import com.architectcoders.equipocinco.extensions.app
 import com.architectcoders.equipocinco.extensions.getViewModel
 import com.architectcoders.equipocinco.framework.SearchManager
+import com.architectcoders.equipocinco.ui.MovieAdapter
+import com.architectcoders.equipocinco.ui.detail.DetailMovieFragment
 import com.architectcoders.generic.framework.extension.isFilled
 import com.architectcoders.generic.framework.extension.view.setVisibleOrGone
 import com.architectcoders.presentation.viewmodels.MovieViewModel
@@ -81,12 +83,13 @@ class MoviesFragment : Fragment() {
     private fun initAdapter(items: List<Movie>) {
         rv?.let {
             rv.layoutManager = GridLayoutManager(activity, 3)
-            adapter = MovieAdapter(items.toMutableList()) {
-                navController.navigate(
-                    R.id.action_moviesFragment_to_detailMovieFragment,
-                    bundleOf(DetailMovieFragment.MOVIE_ID_KEY to it.id)
-                )
-            }
+            adapter =
+                MovieAdapter(items.toMutableList()) {
+                    navController.navigate(
+                        R.id.action_moviesFragment_to_detailMovieFragment,
+                        bundleOf(DetailMovieFragment.MOVIE_ID_KEY to it.id)
+                    )
+                }
             rv.adapter = adapter
             pb.hide()
         }
