+++ data/build.gradle
@@ -4,4 +4,5 @@ apply plugin: 'kotlin'
 dependencies {
     implementation project(":domain")
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation coroutinesDependencies.values()
 }
