+++ usecases/src/main/java/com/gabriel/usecases/GetPopularMoviesUseCase.kt
@@ -8,6 +8,6 @@ import com.architectcoders.domain.state.DataState
 class GetPopularMoviesUseCase(private val moviesRepository: MoviesRepository) : MoviesBaseUseCase<List<Movie>, Any>(){
 
     override suspend fun useCaseExecution(params: Any?): DataState<List<Movie>> {
-       return moviesRepository.getPopularMovies()
+        return moviesRepository.getPopularMovies()
     }
 }
